Вынос функций в сервисный слой Users
