+++ users/urls.py
@@ -1,7 +1,8 @@
 from django.contrib.auth.views import LogoutView
 from django.urls import path
 
-from users.views import RegisterView, email_verification, CustomLoginView, CustomUserUpdate, CustomPasswordResetView, \
+from users.services import UserService
+from users.views import RegisterView, CustomLoginView, CustomUserUpdate, CustomPasswordResetView, \
     CustomPasswordResetDoneView, CustomPasswordResetConfirmView, CustomPasswordResetCompleteView, CustomUserDetailView
 
 app_name = 'users'
@@ -10,7 +11,7 @@ urlpatterns = [
     path('register/', RegisterView.as_view(), name='register'),
     path('login/', CustomLoginView.as_view(), name='login'),
     path('logout/', LogoutView.as_view(), name='logout'),
-    path('confirm/<str:token>/', email_verification, name='confirm'),
+    path('confirm/<str:token>/', UserService.email_verification, name='confirm'),
     path('user/update/<int:pk>/', CustomUserUpdate.as_view(), name='user_update'),
     path('user/profile/<int:pk>/', CustomUserDetailView.as_view(), name='user_profile'),
